[Cast, codehealth] Migrate bitrate storage types from int to uint32_t
Migrate bitrate storage types across media/cast and mirroring service
host from int to uint32_t and fix pre-existing unsafe type conversions.
Also add underflow guards to unsigned bitrate subtraction arithmetic.

Fixed: 216175135
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7981348
Commit-Queue: Jordan Bayles <jophba@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1652036}

+++ media/cast/encoding/media_video_encoder_wrapper_unittest.cc
@@ -248,7 +248,7 @@ TEST_F(MediaVideoEncoderWrapperTest, CanGenerateKeyFrame) {
 }
 
 TEST_F(MediaVideoEncoderWrapperTest, CanSetBitRate) {
-  constexpr int kNewBitRate = 1234567;
+  constexpr uint32_t kNewBitRate = 1234567;
   ExpectEncoderInitialized();
 
   const FrameInfo frame_info = CreateFrameInfo(FrameType::kKey);
@@ -267,9 +267,7 @@ TEST_F(MediaVideoEncoderWrapperTest, CanSetBitRate) {
       .WillOnce([&](const media::VideoEncoder::Options& options,
                     media::VideoEncoder::OutputCB output,
                     media::VideoEncoder::EncoderStatusCB done_cb) {
-        EXPECT_EQ(options.bitrate,
-                  Bitrate::ConstantBitrate(
-                      base::checked_cast<uint32_t>(kNewBitRate)));
+        EXPECT_EQ(options.bitrate, Bitrate::ConstantBitrate(kNewBitRate));
         std::move(done_cb).Run(EncoderStatus::Codes::kOk);
         std::move(quit_closure).Run();
       });
@@ -361,9 +359,9 @@ TEST_F(MediaVideoEncoderWrapperTest, ReportsErrorOnEncodeFailure) {
 // the SetBitRate method without crashing. For motivation, see
 // crbug.com/457353867.
 TEST_F(MediaVideoEncoderWrapperTest, CanHandleMultiplePendingUpdates) {
-  constexpr int kNewBitRate1 = 1234567;
-  constexpr int kNewBitRate2 = 2345678;
-  constexpr int kNewBitRate3 = 3456789;
+  constexpr uint32_t kNewBitRate1 = 1234567;
+  constexpr uint32_t kNewBitRate2 = 2345678;
+  constexpr uint32_t kNewBitRate3 = 3456789;
   ExpectEncoderInitialized();
 
   const FrameInfo frame_info = CreateFrameInfo(FrameType::kKey);
@@ -385,9 +383,7 @@ TEST_F(MediaVideoEncoderWrapperTest, CanHandleMultiplePendingUpdates) {
       .WillRepeatedly([&](const media::VideoEncoder::Options& options,
                           media::VideoEncoder::OutputCB output,
                           media::VideoEncoder::EncoderStatusCB done_cb) {
-        if (options.bitrate ==
-            Bitrate::ConstantBitrate(
-                base::checked_cast<uint32_t>(kNewBitRate3))) {
+        if (options.bitrate == Bitrate::ConstantBitrate(kNewBitRate3)) {
           std::move(quit_closure).Run();
         }
         std::move(done_cb).Run(EncoderStatus::Codes::kOk);
